+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added example for vertical orientation functionality for `terra-tabs`.
+  * Added documentation for FlowsheetDataGrid in `terra-data-grid`.
 
 ## 1.39.0 - (October 3, 2023)